Move cwt parsing to sub module
